Fix Netlify build issues: resolve linting errors and file extensions
- Fix duplicate AIInsights interface declaration in AIPerformanceDashboard.tsx
- Rename lazyLoading.ts to lazyLoading.tsx to support JSX syntax
- Add missing React import to lazyLoading component
- Remove duplicate severity property in ErrorReport interface
- All critical build errors resolved, build now passes successfully
