+++ src/components/AIPerformanceDashboard.tsx
@@ -36,14 +36,7 @@ interface ErrorReport {
   };
   aiPredictedImpact?: number;
   resolutionSuggestions?: string[];
-  severity?: string;
   [key: string]: unknown;
-}
-
-interface AIInsights {
-  predictedHighRiskActions: string[];
-  recommendedImprovements: string[];
-  errorTrends: Array<{ category: string; trend: string }>;
 }
 
 const AIPerformanceDashboard: React.FC<AIPerformanceDashboardProps> = ({ isVisible, onClose }) => {
